Allowing dashes in feed slug / id / user id validations
The current implementation of the validation checks are converting the
dashes to underscores, I assume this is to enable easier checking with
the regexp. This is however causing issues when the user ID is a UUID.
This commit allows the dashes to be considered valid, and moves the
regexp into a global var, as it currently uses the same validation.